feat(deployment):containerized deployment with nginx proxy
