+++ TeaShop/TeaShop/.dockerignore
@@ -1,0 +1,7 @@
+bin/
+obj/
+out/
+.git/
+.vs/
+.vscode/
+infra/certs/
