+++ TeaShop/TeaShop/Deployment/certs/teashop.crt
@@ -1,0 +1,19 @@
+-----BEGIN CERTIFICATE-----
+MIIDCTCCAfGgAwIBAgIUU7QW/TaCSBBNItja/Dxx3rSjg0wwDQYJKoZIhvcNAQEL
+BQAwFDESMBAGA1UEAwwJbG9jYWxob3N0MB4XDTI2MDYwNTE3MDU0NVoXDTI3MDYw
+NTE3MDU0NVowFDESMBAGA1UEAwwJbG9jYWxob3N0MIIBIjANBgkqhkiG9w0BAQEF
+AAOCAQ8AMIIBCgKCAQEAs13BiQUURQVC7MrqWl0jtZSg0e8ZmSfu94RPczBGAKY9
+sv241GLgQB4DcjUHYiUvTsPIotJXUgH6VAusXOJFWMAxDenxwmUH1cPJbKQXOTnn
+crXj7DVYjHv3rt7/GCr/P0wzPYe3X5oNBEiwVRL9lBNTQe1/l2/ihCxsXOcZ6X6s
+H0Lls99d8w69W+Xdp6bTGXSaqHfkGShimoxGVdcbYrfLGhh744kzBsscVwVbzs7e
+h1HohAGQKjbTwA9PSZ5W53YT3/4NDLi9ppFyA9BvpQ/fslT95xqR0VpPNu07Spix
+4PGCWLK/ELc9ayDALdhhOF7yd5k4Fm/tcWONxC7xUQIDAQABo1MwUTAdBgNVHQ4E
+FgQUKFvSR14g0u+hxqU+JTjp86TA0eMwHwYDVR0jBBgwFoAUKFvSR14g0u+hxqU+
+JTjp86TA0eMwDwYDVR0TAQH/BAUwAwEB/zANBgkqhkiG9w0BAQsFAAOCAQEAgrtl
+9JsH9zknm3QNCAXjsEBG0yPc7Hc2RCTqPkcEVZPjlMLO8siUz72s4gBPsbNdlzaT
+A6+Xo9JvjmAZwM900bJ2OqRbBuznjymRUFRhs47fveTlED0LVX13qhUe8G56O8QF
+EWjEQNZwnHQY6oLFEC618wUPhhwYOSU8QsLigaU42dJRUCiGHogV6XKbvZyk2BC6
+NjAO2ZdHWyVGdWAywbAQ5QO6P2rBuZlOCmLE140K0Ss4MZoFQxvxcgo7T1PbOEO3
+sHVM/9/vFzY9SO+q90Stc8bKFuDeKqzp5t5CO3ND38Rms/wOIu4q0RHCImBBY+O5
+QW1GOrKLFFGuX+WZOQ==
+-----END CERTIFICATE-----
