+++ TeaShop/TeaShop/Deployment/certs/teashop.key
@@ -1,0 +1,28 @@
+-----BEGIN PRIVATE KEY-----
+MIIEvQIBADANBgkqhkiG9w0BAQEFAASCBKcwggSjAgEAAoIBAQCzXcGJBRRFBULs
+yupaXSO1lKDR7xmZJ+73hE9zMEYApj2y/bjUYuBAHgNyNQdiJS9Ow8ii0ldSAfpU
+C6xc4kVYwDEN6fHCZQfVw8lspBc5OedytePsNViMe/eu3v8YKv8/TDM9h7dfmg0E
+SLBVEv2UE1NB7X+Xb+KELGxc5xnpfqwfQuWz313zDr1b5d2nptMZdJqod+QZKGKa
+jEZV1xtit8saGHvjiTMGyxxXBVvOzt6HUeiEAZAqNtPAD09JnlbndhPf/g0MuL2m
+kXID0G+lD9+yVP3nGpHRWk827TtKmLHg8YJYsr8Qtz1rIMAt2GE4XvJ3mTgWb+1x
+Y43ELvFRAgMBAAECggEATqsoXzhjE4KjP8gf0VJYfyOZBXqg95OUM+tMDXF2woIn
+PYIatsxHHtgLGuQ3QMyeWyVzsjYfJlDHwmT8TCDR2YMyALBC4Mf/D7q3dXbi2K6U
+3LL6LHeZRsR4dbJIoAj7un9pWetUehHO87oN3MJfhbqUIa5Q79Kavu3c+UJDEy1G
+tcnNzKFfGAaLHBIvg+B+CGYeqlN+N5gl6KTalyAOTrL+zGEuJxB94T9rpg82Ag0k
+O9K/FHnX6ur7gSWdK+MAAsA10lKhyBSHvSsJEeCIWd+kNyqEyLw46prIQX0KLVKi
+CMoWED87rBwBo/7o38HGXT12wyQsZylBRcjx4zmOIQKBgQDucRTPMroH26N+Uhmi
+xlDaq4Uaw0XvLfFg8LTqtGz89VT4DZLHLX1gP4DAHTbk+E/UZL6z6Gz/guOmXfkj
+ZeglwuEBRkfA8Mh5Ycavj4hhUeJIFYZ1cvtzjh0zherHYfi8Mv83Mser53sMDA1R
+B+dgrWUJKraa3Llh7oVXVsk1TwKBgQDAkweCiGuyoeDHvjnZ7TwEv2h9ZcgTLCNK
+5qnxbt/9F/IzdIwGSi+wa6/9ajsVbajzHOJWh40LchNaBRiJpANSFIyWvvi8lpMA
+oz6OU7TDWU492TzdEbeBqpbJN7IM7pElPmdwLVowfHUkMK//Xo+xG7iuqA8Bup/i
+OwdTmkEHXwKBgASJdEro/S8vF7JMn55ARcOTZen3czo01pXwnpCktDov9NjvcS6H
+CA6vxqRhKe07aFFzRUmpD4HhHLHKsfXhd4RYa+2ZR749FqNwn9YxVff8FeEnP7K2
+aGBvgNC9KJV4oMH7+m0zcycrNWUqSYqUKuvlALom7OiPcJuTbUgXJ6bzAoGBAIwA
+GByYyxsxvnXYYLyDUPfuo6tOfJcbpojEitYgEi0w5H7uAwJrmqH6mBZ2zX91mrcr
+z4yyvSKOC22NeNpP564bjF3ndnO8UCDNjSPxzN+cJFqofSwGvk9BUp/0orfuFI1s
+Ac0223OWd58Y2kqgV3ydt7UaLtbF84iH+3qXESGnAoGAb1xS2ysY7OCmBNJ6S4FY
+EtLtr7elXU5qzKR4uryo6b2faT0UDO6Wa1O17VZiDItJPa1JDMFJ8U6k+HSdvH6V
+rp6cE3xZlcD58pqufCzvH9NrM3wU4JMtA7MznxkKln9kfGaUTsXcSg/2YSfVCuOa
+1l1AumQtZxxFLpaoIgqbHNg=
+-----END PRIVATE KEY-----
