+++ TeaShop/TeaShop/Deployment/docker-compose.yml
@@ -1,0 +1,34 @@
+
+networks:
+  teashop-network:
+    driver: bridge
+
+services:
+  teashop-api:
+    container_name: teashop-api
+    build:
+      context: ../
+      dockerfile: Dockerfile    
+    environment:
+      - ASPNETCORE_ENVIRONMENT=Production
+      - ConnectionStrings__DefaultConnection=${DB_CONNECTION_STRING}
+      - Seed__AdminEmail=${SEED_EMAIL}
+      - Seed__AdminPassword=${SEED_PASSWORD}
+    expose:
+      - "8080"
+    networks:
+      - teashop-network
+
+  reverse-proxy:
+    container_name: reverse-proxy
+    image: nginx:alpine
+    ports:
+      - "80:80"
+      - "443:443"
+    volumes:
+      - ./nginx.conf:/etc/nginx/nginx.conf:ro
+      - ./certs:/etc/nginx/certs:ro
+    depends_on:
+      - teashop-api
+    networks:
+      - teashop-network
