+++ TeaShop/TeaShop/Deployment/nginx.conf
@@ -1,0 +1,40 @@
+events {
+    worker_connections 1024;
+}
+
+http {
+    upstream teashop_backend {
+        server teashop-api:8080;
+    }
+
+    server {
+        listen 80;
+        server_name localhost;
+
+        location / {
+            return 400 '{"error": "HTTPS is required."}';
+            default_type application/json;
+        }
+    }
+
+    server {
+        listen 443 ssl;
+        server_name localhost;
+
+        ssl_certificate /etc/nginx/certs/teashop.crt;
+        ssl_certificate_key /etc/nginx/certs/teashop.key;
+
+        ssl_protocols TLSv1.2 TLSv1.3;
+        ssl_ciphers HIGH:!aNULL:!MD5;
+
+        location / {
+            proxy_pass http://teashop_backend;
+            proxy_set_header Host $host;
+            proxy_set_header X-Real-IP $remote_addr;
+            proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+            proxy_set_header X-Forwarded-Proto $scheme;
+
+            proxy_redirect off;
+        }
+    }
+}
