--- TeaShop/TeaShop/Infrastructure/Data/TeaConfiguration.cs
@@ -1,32 +1,0 @@
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata.Builders;
-using TeaShop.Domain.Catalog;
-using TeaShop.Domain.Products;
-
-namespace TeaShop.Infrastructure.Data;
-
-public sealed class TeaConfiguration : IEntityTypeConfiguration<Tea>
-{
-    public void Configure(EntityTypeBuilder<Tea> builder)
-    {
-        builder.ToTable("Teas");
-
-        builder.HasKey(t => t.Id);
-
-        builder.Property(t => t.Name)
-            .IsRequired()
-            .HasMaxLength(150);
-
-        builder.Property(t => t.Price)
-            .IsRequired()
-            .HasPrecision(10, 2);
-
-        builder.Property(t => t.Stock)
-            .IsRequired();
-
-        builder.HasOne<Category>()
-            .WithMany()
-            .HasForeignKey(t => t.CategoryId)
-            .OnDelete(DeleteBehavior.Restrict);
-    }
-}
