+++ TeaShop/TeaShop/Program.cs
@@ -1,8 +1,9 @@
 
+using Microsoft.AspNetCore.HttpOverrides;
 using Microsoft.AspNetCore.Identity;
+using System.Net;
 using TeaShop.Application;
 using TeaShop.Infrastructure;
-using TeaShop.Infrastructure.Data;
 using TeaShop.Infrastructure.Middleware;
 using TeaShop.Infrastructure.Persistence;
 using TeaShop.Infrastructure.Persistence.Seed;
@@ -15,6 +16,16 @@ builder.Services.AddOpenApi();
 
 builder.Services.AddInfrastructure(builder.Configuration);
 builder.Services.AddApplication();
+
+if (builder.Environment.IsProduction())
+{
+    builder.Services.Configure<ForwardedHeadersOptions>(options =>
+    {
+        options.ForwardedHeaders = ForwardedHeaders.XForwardedFor | ForwardedHeaders.XForwardedProto;
+
+        options.KnownIPNetworks.Add(new System.Net.IPNetwork(IPAddress.Parse("172.16.0.0"), 12));
+    });
+}
 
 
 builder.Services.AddTeaShopRateLimiting();
@@ -73,7 +84,10 @@ else
     app.UseHttpsRedirection();
 }
 
-
+if (app.Environment.IsProduction())
+{
+    app.UseForwardedHeaders();
+}
